Modelling/Deploy/hotel-review/requirements.txt
Flask
dill
pandas
google-cloud-bigquery==3.11.0
db-dtypes
gunicorn==20.1.0